asdf-parser/src/main/java/edu/caltech/ipac/roman/asdfparser/AsdfReader.java
package edu.caltech.ipac.roman.asdfparser;

import java.io.*;
import nom.tam.fits.*;
import java.nio.charset.StandardCharsets;
import org.yaml.snakeyaml.Yaml;
import java.util.Map;

public class AsdfReader {

  // Other methods and constructor

    public String readYamlHeader(InputStream inputStream) throws Exception {
        // Implement logic to read and return the YAML header from the input stream
        try (BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))) {
            StringBuilder header = new StringBuilder();
            String line;
            while ((line = reader.readLine()) != null && !line.startsWith("%YAML")) {
                header.append(line).append("\n");
            }
            return header.toString();
        } catch (Exception e) {
            e.printStackTrace();
            return null;
        }
    }
    public void readBinaryBlocks(InputStream inputStream) throws Exception {
            // Assuming we know the offset and structure of the binary blocks
            // int offset = ...;

            try {
                throw new Exception("Not implemented");// Skip to the beginning of the binary block
                // inputStream.skip(offset);

                // Read and process the binary data
                // This will depend on the structure of your binary data
                // e.g., reading into a byte array, processing image data, etc.
            } catch (Exception e) {
                e.printStackTrace();
            }
        }

        public void processFitsBinaryBlock(InputStream inputStream) throws FitsException, IOException {
                // Assuming the FITS binary block is at the current position of inputStream

                // Use Fits class from nom.tam.fits to read the FITS data
                Fits fits = new Fits(inputStream);

                try {
                    // Read all HDUs (Header/Data Units) from the FITS file
                    BasicHDU<?>[] hdus = fits.read();
                    for (BasicHDU<?> hdu : hdus) {
                        // Process each HDU as needed
                        // For example, you can get headers and data, and handle them accordingly
                        Header header = hdu.getHeader();
                        Data data = hdu.getData();

                        // ...process header and data...
                    }
                } finally {
                    // Make sure to close the Fits object to release resources
                    fits.close();
                }
            }


    public static void main(String[] args) {
            if (args.length != 1) {
                System.out.println("Usage: java AsdfReader <path_to_asdf_file>");
                System.exit(1);
            }

            String filePath = args[0];

            try (InputStream inputStream = new FileInputStream(filePath)) {
                AsdfReader reader = new AsdfReader();
                String yamlHeader = reader.readYamlHeader(inputStream);
                System.out.println("YAML Header:\n" + yamlHeader);
            } catch (Exception e) {
                e.printStackTrace();
            }
        }
}
